feat(units): migrate fee/display/policy layer from nanoBTH to picocredits end-to-end
Implements the #649 decision: picocredits (10^-12 BTH) become the only
unit below the UI edge; the two-tier nanoBTH system is retired. Every
migrated value is identical in BTH terms — genuinely nanoBTH-denominated
values scale x1000, mislabeled-but-already-picocredit values are
relabeled only (scaling those would have changed fee verdicts).

- cluster-tax MonetaryPolicy::default()/bitcoin_like() initial_reward:
  50e9 nanoBTH -> 50e12 picocredits (still exactly 50 BTH)
- New MonetaryPolicy::calculate_tail_reward_u128 (identical formula,
  u128 input); node's block.rs tail-reward now delegates to it
- emission_sweep: phase1_supply_nanobth -> phase1_supply_picocredits;
  tail-reward analytics use the u128 entry point (the old u64 clamp
  would silently corrupt results at pico scale); BTH-term sweep goldens
  pass unchanged
- tokens::Bth::MINIMUM_FEE: 400 microBTH now in picocredits
  (400_000 -> 400_000_000; FeeMap-internal only)
- transaction/types/constants.rs: two-tier contract and nanoBTH
  constants deleted; single-unit doc
- Relabels (values unchanged, already picocredits): dynamic_fee,
  mempool/RPC docs (baseRate, cluster wealth), wallet fee estimation,
  validation error docs
- balance CLI prints formatted BTH + picocredits (nanoBTH line removed)
- web docs.tsx: smallest-unit table corrected to picocredit/12 decimals
- u128/i128 widening of three sim-only multiply sites that pico-scale
  values overflow (tag mix, transfer mass deltas, wash-trade savings)
- PROTOCOL_VERSION 4.0.0 -> 4.1.0 (MINOR: RPC unit declarations change,
  no consensus rule does; MIN_SUPPORTED stays 4.0.0, 4.0.0 peers warn
  and stay connected — pinned by a new regression test)

Fixes #694

+++ cluster-tax/src/tag.rs
@@ -114,7 +114,8 @@ impl TagVector {
     /// - `incoming`: tag vector of incoming coins
     /// - `incoming_value`: value of incoming coins
     pub fn mix(&mut self, self_value: u64, incoming: &TagVector, incoming_value: u64) {
-        let total_value = self_value + incoming_value;
+        // u128: two u64 values can sum past u64::MAX.
+        let total_value = self_value as u128 + incoming_value as u128;
         if total_value == 0 {
             return;
         }
@@ -127,14 +128,22 @@ impl TagVector {
             }
         }
 
-        // Compute weighted average for each cluster
+        // Compute weighted average for each cluster.
+        //
+        // Widened to u128 (#694): at picocredit scale a value can exceed
+        // u64::MAX / TAG_WEIGHT_SCALE (~1.84e13 picocredits ≈ 18.4 BTH), so
+        // `value * weight` must not be computed in u64. The result is a
+        // weighted average of weights (<= TAG_WEIGHT_SCALE), so it always fits
+        // TagWeight; the wider intermediate changes no result that did not
+        // previously overflow.
         for cluster in all_clusters {
-            let self_weight = self.get(cluster) as u64;
-            let incoming_weight = incoming.get(cluster) as u64;
+            let self_weight = self.get(cluster) as u128;
+            let incoming_weight = incoming.get(cluster) as u128;
 
             // new_weight = (self_value * self_weight + incoming_value * incoming_weight) /
             // total_value
-            let numerator = self_value * self_weight + incoming_value * incoming_weight;
+            let numerator =
+                self_value as u128 * self_weight + incoming_value as u128 * incoming_weight;
             let new_weight = (numerator / total_value) as TagWeight;
 
             self.set(cluster, new_weight);
